+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.types.js
@@ -21,7 +21,7 @@ export type PlainProps = {|
     onGoBack: () => void,
     onEnrollmentError: (message: string) => void,
     onEnrollmentSuccess: () => void,
-    onCancelEditEvent: () => void,
+    onCancelEditEvent: (isScheduled: boolean) => void,
     onHandleScheduleSave: (eventData: Object) => void,
     pageStatus: string,
     eventStatus?: string,
@@ -35,4 +35,5 @@ export type Props = {|
     enrollmentId: string,
     orgUnitId: string,
     eventId: string,
+    initMode?: string,
 |};