+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/programStage/ProgramStageFactory.js
@@ -65,7 +65,7 @@ export class ProgramStageFactory {
                             { sectionDataElement }));
                     return;
                 }
-                const element = await this.dataElementFactory.build(cachedProgramStageDataElement);
+                const element = await this.dataElementFactory.build(cachedProgramStageDataElement, section);
                 element && section.addElement(element);
             });
         }
@@ -81,7 +81,7 @@ export class ProgramStageFactory {
         if (cachedProgramStageDataElements) {
             // $FlowFixMe
             await cachedProgramStageDataElements.asyncForEach((async (cachedProgramStageDataElement) => {
-                const element = await this.dataElementFactory.build(cachedProgramStageDataElement);
+                const element = await this.dataElementFactory.build(cachedProgramStageDataElement, section);
                 element && section.addElement(element);
             }));
         }
@@ -128,6 +128,7 @@ export class ProgramStageFactory {
             _stage.generatedByEnrollmentDate = !!cachedProgramStage.generatedByEnrollmentDate;
             _stage.reportDateToUse = cachedProgramStage.reportDateToUse;
             _stage.minDaysFromStart = cachedProgramStage.minDaysFromStart;
+            _stage.repeatable = cachedProgramStage.repeatable;
             _stage.stageForm = new RenderFoundation((_form) => {
                 _form.id = cachedProgramStage.id;
                 _form.name = cachedProgramStage.displayName;