+++ src/core_modules/capture-core/metaDataStoreLoaders/programs/quickStoreOperations/storePrograms.js
@@ -80,7 +80,7 @@ const convert = (() => {
     };
 })();
 
-const fieldsParam = 'id,displayName,displayShortName,description,programType,style,displayFrontPageList,onlyEnrollOnce,' +
+const fieldsParam = 'id,displayName,displayShortName,description,programType,style,displayFrontPageList,useFirstStageDuringRegistration,onlyEnrollOnce,' +
 'minAttributesRequiredToSearch,enrollmentDateLabel,incidentDateLabel,' +
 'featureType,selectEnrollmentDatesInFuture,selectIncidentDatesInFuture,displayIncidentDate,' +
 'dataEntryForm[id,htmlCode],' +